requirement.txt
Flask==1.0.2
Flask-Admin==1.5.1
Flask-DebugToolbar==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Jinja2==2.10
MySQL-python==1.2.5
pyldap==2.4.45
Flask-SimpleLDAP==1.2.0
